+++ includes/core/index.php
@@ -4,11 +4,20 @@
 defined( 'ABSPATH' ) or exit;
 
 
+// API
+require_once 'api/api-events.php';
+require_once 'api/api-locations.php';
+require_once 'api/api-partners.php';
+require_once 'api/api-sessions.php';
+require_once 'api/api-speakers.php';
+
+
 // Custom Post Types
 require_once 'post-types/post-type-speaker.php';
 require_once 'post-types/post-type-session.php';
 require_once 'post-types/post-type-partner.php';
 require_once 'post-types/post-type-exhibition-space.php';
+
 
 // Custom Taxonomies
 require_once 'taxonomies/taxonomy-partnership.php';