+++ includes/shortcodes/index.php
@@ -1,0 +1,11 @@
+<?php
+/** Prevent direct access */
+
+defined( 'ABSPATH' ) or exit;
+
+
+// Shortcode Classes
+require_once 'class-shortcode-icon-wall.php';
+require_once 'class-shortcode-event-table.php';
+require_once 'class-shortcode-speaker-grid.php';
+require_once 'class-shortcode-exhibition-list.php';